+++ R/pddply.R
@@ -5,8 +5,7 @@
 ##' All of the arguments except \code{njobs} are passed directly to arguments of the same name in
 ##' \code{\link[plyr:ddply]{plyr::ddply}}.
 ##'
-##' An innocuous warning is thrown when \code{\link[plyr:ddply]{plyr::ddply}} is called in parallel:
-##' \url{https://github.com/hadley/plyr/issues/203}.  This function catches and hides that warning, which looks like this:
+##' An innocuous warning is thrown when \code{\link[plyr:ddply]{plyr::ddply}} is called in parallel.  This function catches and hides that warning, which looks like this:
 ##' \verb{
 ##'
 ##' Warning messages: